requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.0.0
anthropic>=0.20.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
elevenlabs>=1.0.0
openai-whisper>=20231117
Pillow>=10.0.0
